Fixed #257, but it doesn't end there I'm afraid
As of PHP7 'reference' is an actual type, meaning that all the is* methods
would return false if it is a reference. That's fixed with this commit, however
the actual zval is of course a reference and not said value/type. Meaning that
besides at this array iterator this didn't work, we now probably have a lot more
locations in our code where a reference will simply not work. This commit however
added a dereference() method to our zval class to at least make this 'easy' to solve.
As this dereference() function will dereference the reference or just return the value,
meaning you can just drop it in there.