Improve performance of destructing objects
We used to throw a NotImplemented exception, but this causes the c++ runtime
to allocate on the heap. We can call zend_objects_destroy_object directly
in the Php::Base::__destroy method instead.

If someone overrides this method, then zend_objects_destroy_object will not
be called, which is the behavior we had before as well.